ROX-20970: Tests for Approved False Positives Table
