+++ ui/apps/platform/cypress/integration/vulnerabilities/exceptionManagement/approveRequestFlow.test.ts
@@ -1,24 +1,11 @@
 import withAuth from '../../../helpers/basicAuth';
 import { hasFeatureFlag } from '../../../helpers/features';
-import { getInputByLabel } from '../../../helpers/formHelpers';
 import { cancelAllCveExceptions } from '../workloadCves/WorkloadCves.helpers';
-import { deferAndVisitRequestDetails } from './ExceptionManagement.helpers';
+import { deferAndVisitRequestDetails, approveRequest } from './ExceptionManagement.helpers';
 
 const comment = 'Defer me';
 const expiry = 'When all CVEs are fixable';
 const scope = 'All images';
-
-export function approveRequest() {
-    cy.get('button:contains("Approve request")').click();
-    cy.get('div[role="dialog"]').should('exist');
-    getInputByLabel('Approval rationale').type('Approved');
-    cy.get('div[role="dialog"] button:contains("Approve")').click();
-    cy.get('div[role="dialog"]').should('not.exist');
-    cy.get('div[aria-label="Success Alert"]').should(
-        'contain',
-        'The vulnerability request was successfully approved.'
-    );
-}
 
 describe('Exception Management Request Details Page', () => {
     withAuth();
